backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
sqlalchemy==2.0.23
pymysql==1.1.0
redis==5.0.1
hiredis==2.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
loguru==0.7.2
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.1
oss2>=2.18.0
aiosmtplib>=2.0.0
pillow>=9.0.0
openpyxl>=3.1.0
gunicorn>=21.2.0

